* Do the same change in the application ecosystem

+++ apps/Calc/src/main.c
@@ -87,7 +87,7 @@ void CopyTextBoxContents(Window* pWindow)
 }
 char syms[] = "0123456789=%*/+-C";
 #define WND_BORDER_SIZE 3
-void CALLBACK WndProc (Window* pWindow, int messageType, int parm1, int parm2)
+void CALLBACK WndProc (Window* pWindow, int messageType, long parm1, long parm2)
 {
 	switch (messageType)
 	{
