+++ apps/Chess/src/main.c
@@ -454,7 +454,7 @@ void ChessUpdateMoveList()
 
 void AddMoveList(Window* pWindow, Rectangle rect, int comboID);
 
-void CALLBACK ChessWndProc (Window* pWindow, int messageType, int parm1, int parm2)
+void CALLBACK ChessWndProc (Window* pWindow, int messageType, long parm1, long parm2)
 {
 	switch (messageType)
 	{
