+++ apps/Chess/src/propopup.c
@@ -13,7 +13,7 @@ extern Window* g_pWindow;
 
 ePiece g_PromotedPiecePlace;
 
-void CALLBACK PromoPopupWndProc (Window* pWindow, int messageType, int parm1, int parm2)
+void CALLBACK PromoPopupWndProc (Window* pWindow, int messageType, long parm1, long parm2)
 {
 	switch (messageType)
 	{
