+++ apps/Doom/src/doomgeneric_ns.c
@@ -301,7 +301,7 @@ int g_nextUpdateIn, timer;
 Image image;
 bool gotHexE0 = false;
 
-void CALLBACK WndProc (Window* pWindow, int messageType, int parm1, int parm2)
+void CALLBACK WndProc (Window* pWindow, int messageType, long parm1, long parm2)
 {
 	switch (messageType)
 	{
