+++ apps/GameTest/src/main.c
@@ -32,7 +32,7 @@ bool IsKeyDown (int keyCode)
 	return gKeyboardState [keyCode];
 }
 
-void CALLBACK WndProc (Window* pWindow, int messageType, int parm1, int parm2)
+void CALLBACK WndProc (Window* pWindow, int messageType, long parm1, long parm2)
 {
 	switch (messageType)
 	{
