+++ apps/VBuilder/src/w_main.c
@@ -122,7 +122,7 @@ void VbSelectControlDialog()
 	RegisterEvent(g_pFormDesignerWindow, EVENT_PAINT, 0, 0);
 }
 
-void CALLBACK PrgVbMainWndProc (Window* pWindow, int messageType, int parm1, int parm2)
+void CALLBACK PrgVbMainWndProc (Window* pWindow, int messageType, long parm1, long parm2)
 {
 	switch (messageType)
 	{
