+++ apps/WavePlayer/src/main.c
@@ -376,7 +376,7 @@ void SetPos(int pos)
 	lseek(g_fileFD, g_offsetFromStart + filePos, SEEK_SET);
 }
 
-void CALLBACK WndProc (Window* pWindow, int msg, int parm1, int parm2)
+void CALLBACK WndProc (Window* pWindow, int msg, long parm1, long parm2)
 {
 	switch (msg)
 	{
