+++ apps/WorldMap/src/main.c
@@ -49,7 +49,7 @@ void SetSelectedCountry(int country)
 	}
 }
 
-void CALLBACK WndProc (Window* pWindow, int messageType, int parm1, int parm2)
+void CALLBACK WndProc (Window* pWindow, int messageType, long parm1, long parm2)
 {
 	switch (messageType)
 	{
